Improved notifications and permissions
Changed it from persist=.runelite to filesystem=~/.runelite:create to
allow for better persistence and ease of access for the user.

Removed the socket=wayland since RuneLite doesn't support it.

Added libnotify support so that notifications, such as the idle
notification, use the proper system notifications and not the horrible
RuneLite fallback ones.